pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-012"
version = "0.1.0"
description = ""
authors = ["Oleksandr Filippov <7philipps7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
fastapi-users = "^14.0.0"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"
redis = {extras = ["asyncio"], version = "^5.2.0"}
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"



